refactor: move Cargo placeholder manifests into monochange_cargo
