+++ .changeset/move-cargo-placeholder-manifest.md
@@ -1,0 +1,8 @@
+---
+monochange: patch
+monochange_cargo: minor
+---
+
+# Move Cargo placeholder manifests into monochange_cargo
+
+Move crates.io placeholder `Cargo.toml` and `src/lib.rs` generation out of `monochange::package_publish` and into `monochange_cargo`.
